Billing service test

+++ src/main/java/com/example/clinic/app/billing/dto/InvoiceDTO.java
@@ -3,11 +3,12 @@ package com.example.clinic.app.billing.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @Builder
 public class InvoiceDTO {
     private List<ConsultationDTO> consultations;
-    private double totalCost;
+    private BigDecimal totalCost;
 }
